# Faster generation of candidates