On my machine the test case with of 124 checks runs roughly a second faster with these changes.