@fsondej 2020-03-25
I did some benchmarks

before changes:
english words        3.397s
english sentences    0.452s
polish words         0.849s

after adding __slots__:
english words        3.458s
english sentences    0.445s
polish words         0.860s

after chain:
english words        3.040s
english sentences    0.386s
polish words         0.765s

after generator expressions:
english words        3.081s
english sentences    0.396s
polish words         0.769s

generator expressions don't give any speedup, but make the code less readable so I reverted them

slots and chains are nice, they speedup about 15%

[ins] In [20]: %timeit -n1 -r1 count_words('autocorrect/data/plwiki.xml', 'pl')
before Counter:
25min 28s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
after Counter:
21min 48s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
and the output is the same

good job :)