@boidolr 2020-03-25
Interesting - for me the generators did make a difference. Though it wasn't as big as the others.
As there was no benefit for you the readability argument certainly wins here :)