@fsondej 2020-03-25
I added some benchmarks to unit tests, you can run them and write how it went. Speed fluctuates randomly, so the benchmarks take only minimum time from many runs so it should be more robust.