@boidolr 2020-03-25
Most of the time is now spent in `str.join` and the [check for existing words](https://github.com/fsondej/autocorrect/blob/64256aaaa924c36bd85c4fe40eaebeeeb82ab8ba/autocorrect/__init__.py#L35) so I do not think there is anything left which is reasonable to change.